+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added : evolution summary switch to toggle evolution mode
 - added : evolution summary filter main entity
+- added : evolution summary filter simple filter mode
 - changed : icon "partie de" use entity color
 - changed : evolution update date use futur and past color for variable icon
 - changed : put tab title on two rows